docs : post-mapping.json 프로퍼티 추가 설정 #117

+++ src/main/java/com/midas/shootpointer/domain/post/dto/response/PostSort.java
@@ -3,7 +3,7 @@ package com.midas.shootpointer.domain.post.dto.response;
 import jakarta.validation.constraints.NotNull;
 
 public record PostSort(
-        @NotNull Double _score,
+        @NotNull float _score,
         @NotNull Long likeCnt,
         @NotNull Long lastPostId
 ) {
